+++ js/main.js
@@ -1,2 +1,3 @@
 import './gallery.js';
 import './form.js';
+import './validate-tags-form.js';
